+++ packages/webview_flutter/android/src/main/java/io/flutter/plugins/webviewflutter/InputAwareWebView.java
@@ -43,14 +43,11 @@ final class InputAwareWebView extends WebView {
 
   /** Sets the proxy adapter view back to its default behavior. */
   void unlockInputConnection() {
-    if (proxyAdapterView != null) {
-      proxyAdapterView.setLocked(false);
+    if (proxyAdapterView == null) {
+      return;
     }
 
-    // Restart the input connection to avoid ViewRootImpl assuming an incorrect window state.
-    InputMethodManager imm =
-        (InputMethodManager) containerView.getContext().getSystemService(INPUT_METHOD_SERVICE);
-    imm.restartInput(containerView);
+    proxyAdapterView.setLocked(false);
   }
 
   /** Restore the original InputConnection, if needed. */