+++ packages/webview_flutter/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 * Add an optional `userAgent` property to set a custom User Agent.
   Add `getUserAgent` to WebViewController. 
+  
+## 0.3.11+2
+
+* Add fix for input connection being dropped after a screen resize on certain
+  Android devices.
+
+## 0.3.11+1
+
+* Work around a bug in old Android WebView versions that was causing a crash
+  when resizing the webview on old devices.
 
 ## 0.3.11